desktop: Setting proper scaling for 32-inch monitor
In the past I've just assumed that Wayland had the same issues with
fractional scaling that Xorg had so I never bothered with it. I just
zoomed in and set high font sizes. No more!

The Emacs font size is still a little too large probably but I'm
playing around with it for now.