Use nonguix substitute servers for all machines.
Usually I had no issues with only having proteus the laptop use the
substitutes. My desktop is more than powerful enough to build the
packages like linux and firefox itself quite quickly. Unfortunately I
encountered a build issue with Firefox 122 that I just was not able to
resolve. So substitutes it is.